Main notebook file with fixed output error (8/11)
In this file, I've fixed the issue with the output errors, exported the pyLDAvis interactive figure that otherwise cannot be seen, and commented out some lines. Other than that I did not modify the contents compared to the original file that I've uploaded before the deadline.
I am also uploading the exported html version of pyLDAvis to show that the figure was, well, real.